Finished Challenge
I have enabled the program to accurately create the Baum-Sweet sequence correctly outputting the numbers from 1 - 20 and have correlated my results against the numbers on https://oeis.org/A086747.

I have also commented my code effectively allowing for quick and easy reading of my solution.

I have not added any kind of error handling for this program (non integer inputs and the like) as it is only me using the project but, it might be a good idea for the future to include these for good practice.